Fix issue 379 - regression with page links
See https://code.google.com/p/daisy-pipeline/issues/detail?id=379

Relative links were not correctly computed when the base URI passed as
parameter was that of a document and not a directory.
Rather than forcing caller to always pass a URI of a directory, the XSLT
now systematically computes the URI of the directory from the URI passed
as the `base-uri` paramater.