pagebreaks can be unnumbered. Use a hyphen to represent them in the navdoc.
Not only can they be unnumbered but they don't have to have an implicit number either.
A typical example is a series of images, one image on each page, where the page numbering
continues on the following page as if there had been no pagebreaks. Like this:

* page #1
* page #2
* page with image
* page with image
* page #3
* page #4

It is also possible for books to skip or repeat page numbers (either by mistake or intentional),
but that's another issue not affected by this commit.