Extract generic chunker step from html-chunker
The generic part is just a dumb, HTML unaware chunker that chunks
based on a (HTML specific) style sheet and preserved the structure of
the input. The generic step is followed by a "finalize" step that
cleans up the structure of the resulting chunks.

See https://github.com/daisy/pipeline-scripts/issues/123