Implement px:chunker in Java
Instead of the full-blown XSLT style sheet support, the break points
are now specified with a simple XSLTMatchPattern. The reason is
because I don't know a good way to implement the XSLT style sheet
support in Java, but also because a simpler solution is preferable as
long as it fulfills the requirements.

Note that the behavior of px:chunker has changed slightly, but the
behavior px:html-chunker has stayed the same.