Add a px:add-xml-base step
which apparently can be used to "fix up" documents that have elements
with an empty base URI. Such documents are created in some places in
our code and are possibly the result of quirks in XMLCalabash.

px:add-xml-base is basically a wrapper around the standard
p:add-xml-base, but it has an additional option to omit the xml:base
attribute on the document element.