Move px:create-mediaoverlays to epub3-utils
and rename it to px:epub3-create-mediaoverlays.

Also, now that mediaoverlay-utils does not contain anything EPUB
specific anymore, rename it to smil-utils.

px:mediaoverlay-rearrange and px:mediaoverlay-join are more or less
generic (although they are @epub:textref aware), and
px:mediaoverlay-upgrade-smil is for converting between the DAISY 2.02,
DAISY 3 and EPUB 3 flavours of SMIL.