Change behavior of px:fileset-copy and px:fileset-move, and add px:fileset-delete
px:fileset-copy now does what px:fileset-move did: "lazily" copy a
fileset to a new location. No px:copy calls are done in this step. The
actual copying is deferred to when px:fileset-store is called on the
resulting fileset.

Two new options were added: "flatten" and "prefix".

px:fileset-move does the same as px:fileset-copy but in addition
deletes the original files on disk. The actual deletion is deferred to
when the new step px:fileset-delete is called on the original fileset.