px:epub3-create-mediaoverlays: don't change the locations of audio files
The d:audio-clips document is assumed to contain the right
information, with the audio files at the correct location within the
EPUB (because px:audio-clips-update-files has been applied after the
px:fileset-copy).

This means that the "audio-dir" option is not needed anymore. Also,
the step does not have to assume anymore that all audio files are
copied to a single directory. We might want to preserve the directory
structure, for example when converting from a DAISY 3 with audio.