New "ace-adapter" module
Add new (Java based) XProc step named "px:ace" to allow the use of a
system wide install of the "Accessibility Checker for EPUB" tool to
create an accessibility compliance report. The presence of Ace on the
PATH of the user is checked when the provider is activated.

Currently the report is loaded from file using px:html-load. An issue
should be created in the Ace github repo to request that the report
would be stored as XTHML.