Optimize px:fileset-store for files that originate from a ZIP
Also make sure that files that originate from a ZIP and are not loaded
in memory are not affected by serialization attributes when copied to
the new location.

px:fileset-store now fails if fileset contains files outside of the
base directory.

The code was significantly simplified, e.g. by making use of
pxi:fileset-fix-original-hrefs.