Create a generic px:smil-to-audio-clips
and use it in daisy3-to-epub3.

Also in daisy3-to-epub3, don't prefix ID refs in NAV and SMIL
documents in case of multiple input DTBooks (to avoid ID collisions),
because this is also not done in the content documents, and because a
much cleaner solution is to preprocess the DTBooks.