Remove default.css from the catalogs
The "default-stylesheet" options have a default value, so it does not
need to be specified and therefore does not need to be publicly
available.

Also remove bana.css and ueb.css from the zedai-to-pef catalog for now
as they are basically just stubs.

Kept reset.css for now (but changed the URL).