Add an experimental html-to-dtbook script
which is just a wrapper around the existing px:html-to-dtbook
step. For now it does not support multiple HTML files. If the input is
actually an EPUB, use the epub3-to-dtbook script. If the input is
DAISY 2.02, use daisy202-to-epub3 and then epub3-to-dtbook. Otherwise
use html-to-epub3 and then epub3-to-dtbook.

For the time being we don't expose the script yet.