Invoke generate-toc.xsl from generate-toc.scss
Note that the behavior changed slightly in the sense that the style
sheet now takes as input the result of any XSLTs supplied through the
"stylesheets" option applied to the source document. Before,
generate-toc.xsl would be applied before any provided XSLTs. If
needed, adapt your XSLT style sheets so that they will not affect the
TOC.