Rename "ascii-table" option to "preview-table"
The table of the ASCII file format can now only be specified within
the ascii-file-format query. This may result in some duplication in
the options the user has to specify, but it simplifies the code and
makes it more clear what the options do.

Fixes https://github.com/daisy/pipeline-mod-braille/issues/163 and
https://github.com/daisy/pipeline-mod-braille/issues/75.