Remove getVersion() from TTSService API
because it serves no purpose and might be confusing.

It is not possible to specify the version in CSS or in a TTS
configuration file. So two services with the same name that serve the
same voices are considered equivalent. The overall priority of the
services determines which one is used. This would only be useful if
services with lower priority would automatically be used as fallback
when the first fails, and this is also what people might
assume. However it is currently not the case.

Because there are currently no TTS services with the same name and
different versions, I decide to just drop the version.