Eliminate concept of AudioBuffer and AudioBuffer allocation
Audio data is now captured in AudioInputStreams and memory does not
need to be allocated explicitly anymore. AudioBufferTracker, which is
renamed to "AudioFootprintMonitor", now only monitors the memory
footprint of the audio data returned by TTSEngine implementations. It
does not control the memory allocation directly anymore, but can still
be used to regulate the memory consumed by TTS and encoding threads by
intervening when the memory exceeds certain thresholds.