Deprecate the support for XSLT braille style sheets
It is now advised to apply any XSLT style sheets during
pre-processing.

The reason for this deprecation is that XSLT is not really a style
sheet language like CSS but rather a fully-fledged programming
language, which means the feature could be used for a wrong purpose
and, more importantly, poses a security issue.

Running code supplied by the user is not only dangerous, it can also
make it unclear who is responsible for unwanted behavior, Pipeline or
the user-supplied code. If the XSLT transforms the document in a
non-trivial way, that makes it hard to tell exactly which document the
CSS is applied to, which makes it hard to debug the style sheets.