Use px:fileset-add-entries instead of p:for-each loop with px:fileset-add-entry
This is a pattern that recurs a lot and is not efficient.

Just to be sure, keep the px:fileset-join that follows the for loop
because of its side effects (which may or may not matter), and also
because px:fileset-add-entries does not remove/join duplicates within
the added entries.