dtbook-to-zedai: Change "assert-valid" option to "validation"
with the three values "abort", "report" and "off". The latter can be
used for instance for when you know the input is a valid NIMAS XML
file (which is not necessarily a valid DTBook).

Also don't report validation issues in intermediary documents (except
when the value of the option is "abort").