Refactor px:block-translate
Inline ::before and ::after pseudo-elements are now taken into account
when pre-translating the containing element.

'content' properties that are not inside a ::before or ::after
pseudo-element, or that are inside a ::before or ::after
pseudo-element stacked onto a different kind of pseudo-element or
other rule, are now not pre-translated anymore.

Fixes issues https://github.com/daisy/pipeline/issues/652 and
https://github.com/daisy/pipeline/issues/603.