px:report-errors: make use of px:log-error and px:error
The code option to specify the error code was removed. The step now
throws the errors on the "report" port if the "method" option is set
to "error".

px:error and px:log-error now accept a sequence of documents on the
"error" port. For convenience, px:error now also has a "source" port.