Update to latest framework
in which the property system was reworked.

Support for the org.daisy.pipeline.tts.host.protection property has
been removed. It was already deprecated. Now it is not possible
anymore to set properties through the XML config file anymore, and
properties that are considered safe are now always marked settable.