dtbook-to-rtf: take into account enum and start attributes on lists
see https://github.com/daisy/pipeline/issues/685

This is a quick hack that preformats lists which have "start" or
"enum" attributes. Other lists are handled as before. The downside of
this solution is that preformatted lists look different from regular
"ol" an "ul" lists, and do not behave the same in MS Word.