dtbook-to-epub3: don't list resources both as link element and manifest item
Resources must not be listed both as a "link" element in the package
metadata and as a manifest item. For simplicity, we assume that when
the metadata contains a link element, the resource is not referenced
anywhere else and can therefore be safely removed from the
manifest. Note that this is a very brittle solution.