Add AWS polly adapter for TTS
The original code was provided by ONCE. The adapter was then updated
for the latest version of Pipeline and refactored to match the
structure of other adapters (notably the Azure and Google ones). The
adapter was also updated to allow for the use of generative,
long-form, or neural voices if available for the user's AWS account.

The main changes compared to original code from ONCE are:

- Updated the service and engine classes to mimic Azure adapter.
- Added the following configuration properties for connecting to AWS:
  - org.daisy.pipeline.tts.aws.accesskey
  - org.daisy.pipeline.tts.aws.secretkey
  - org.daisy.pipeline.tts.aws.region
- Omit namespace prefixes from SSML as AWS seems to refuse it.
- Cleaned up tests files