Handle new sidebar styles from SaveAsDAISY
Handle four new styles required to create sidebar elements:

- Sidebar - Optional (DAISY)
- Sidebar - Required (DAISY)
- Sidebar header - Optional (DAISY)
- Sidebar header - Required (DAISY)

The previous way to create a sidebar was to use a textbox, but I was
not able to produce the expected structure with Word (it seems the
underlying structure in Word has changed).

Also fixes an issue with lists within blocks like poem, epigraph,
sidebar and prodnote, where the block was closed before the list
element (causing invalid XML).